feat: add demo mode

+++ src/services/mongoose/store/dish.dal.ts
@@ -1,7 +1,9 @@
+import { validateRequest } from "@/lib/lucia/auth";
 import { cloudinaryIdentifier } from "@/lib/utils/cloudinary";
 import { Dish, DishDoc, TDish } from "@/schemas/mongoose/store/dish.model";
 import { MenuSchema } from "@/schemas/zod/store/menu.schema";
 import { deleteImage, uploadImage } from "@/services/api/cloudinary";
+import { Types } from "mongoose";
 import { revalidateTag, unstable_cache } from "next/cache";
 import slugify from "slugify";
 import { validateRole } from "../guard";
@@ -42,84 +44,141 @@ interface GetDishesResponse {
   currentPage: number;
 }
 
-const getDishes = unstable_cache(
-  async ({ course, featured, popular, page = 1, limit = PAGE_LIMIT }: GetDishesArg): Promise<GetDishesResponse> => {
-    try {
-      await dbConnect();
+const getDishes = async ({
+  course,
+  featured,
+  popular,
+  page = 1,
+  limit = PAGE_LIMIT,
+}: GetDishesArg): Promise<GetDishesResponse> => {
+  const { user } = await validateRequest();
+  const userId = user?.role === "demo" ? user.id : undefined;
 
-      // Build the filter
-      const filter: Record<string, any> = {};
-      if (course) filter.course = course;
-      if (featured !== undefined) filter.featured = featured;
-      if (popular !== undefined) filter.popular = popular;
+  return unstable_cache(
+    async () => {
+      try {
+        await dbConnect();
 
-      // Pagination options
-      const options = { limit, skip: (page - 1) * limit, sort: "-createdAt" };
+        // Build the filter
+        const filter: Record<string, any> = { userId };
+        if (course) filter.course = course;
+        if (featured !== undefined) filter.featured = featured;
+        if (popular !== undefined) filter.popular = popular;
 
-      // Get total number of documents for pagination
-      const totalDishes = await Dish.countDocuments(filter).exec();
-      const totalPages = Math.ceil(totalDishes / limit);
+        // Pagination options
+        const options = { limit, skip: (page - 1) * limit, sort: "-createdAt" };
 
-      // Handle the case where the requested page is greater than the total number of pages
-      if (page > totalPages) {
-        return {
-          dishes: [],
-          totalPages,
-          currentPage: totalPages,
+        // Get total number of documents for pagination
+        const totalDishes = await Dish.countDocuments(filter).exec();
+        const totalPages = Math.ceil(totalDishes / limit);
+
+        // Handle the case where the requested page is greater than the total number of pages
+        if (page > totalPages) {
+          return {
+            dishes: [],
+            totalPages,
+            currentPage: totalPages,
+          };
+        }
+
+        // Fetch the dishes for the current page
+        const dishes = await Dish.find(filter, {}, options).exec();
+
+        // Fetch special diets for the restaurant
+        const existingSpecialDiets = await getSpecialDiets();
+
+        // Create DishDTO
+        const dishesDTO = dishes.map((dish) => createDishDTO(dish, existingSpecialDiets));
+
+        return { dishes: dishesDTO, totalPages, currentPage: page };
+      } catch (error) {
+        console.error(error);
+        throw new Error("Failed to fetch dishes.");
+      }
+    },
+    [
+      "dishes",
+      course ?? "",
+      featured?.toString() ?? "",
+      popular?.toString() ?? "",
+      page.toString(),
+      limit.toString(),
+      userId ?? "",
+    ],
+    { tags: ["dishes"], revalidate: 60 * 60 }
+  )();
+};
+
+const getDish = async ({ id, slug }: { id?: string; slug?: string }): Promise<DishDTO> => {
+  const { user } = await validateRequest();
+  const userId = user?.role === "demo" ? user.id : undefined;
+
+  return unstable_cache(
+    async () => {
+      try {
+        await dbConnect();
+        const dish = await Dish.findOne({ userId, $or: [{ _id: id }, { slug }] }).exec();
+
+        if (!dish) {
+          throw new Error("Dish not found.");
+        }
+
+        // Fetch special dietary options for the restaurant
+        const existingSpecialDiets = await getSpecialDiets();
+
+        const dishDTO = createDishDTO(dish, existingSpecialDiets);
+
+        return dishDTO;
+      } catch (error) {
+        console.error(error);
+        throw new Error("Failed to fetch dish.");
+      }
+    },
+    ["dish", id ?? slug ?? "", userId ?? ""],
+    { tags: ["dish"], revalidate: 60 * 60 * 24 }
+  )();
+};
+
+const createDemoDishes = async (userId: string) => {
+  try {
+    await dbConnect();
+
+    const dishes = await Dish.find({ userId: undefined }).exec();
+
+    const dishesDemo = await Promise.all(
+      dishes.map(async (dish) => {
+        const imageUrl = dish.image.imageUrl;
+        const uploadResult = await uploadImage(imageUrl, {
+          folder: `/food-ordering-app/restaurant-demo/menu/${dish.course}`,
+        });
+
+        const dishDemo = {
+          ...dish.toObject(),
+          _id: new Types.ObjectId(),
+          image: cloudinaryIdentifier(uploadResult) as any,
+          userId,
+          expiresAt: new Date(Date.now() + 1000 * 60 * 60), // 1 hour
         };
-      }
 
-      // Fetch the dishes for the current page
-      const dishes = await Dish.find(filter, {}, options).exec();
+        return dishDemo;
+      })
+    );
 
-      // Fetch special diets for the restaurant
-      const existingSpecialDiets = await getSpecialDiets();
-
-      // Create DishDTO
-      const dishesDTO = dishes.map((dish) => createDishDTO(dish, existingSpecialDiets));
-
-      return { dishes: dishesDTO, totalPages, currentPage: page };
-    } catch (error) {
-      console.error(error);
-      throw new Error("Failed to fetch dishes.");
-    }
-  },
-  ["dishes"],
-  { tags: ["dishes"], revalidate: 60 * 60 }
-);
-
-const getDish = unstable_cache(
-  async ({ id, slug }: { id?: string; slug?: string }): Promise<DishDTO> => {
-    try {
-      await dbConnect();
-      const dish = await Dish.findOne({ $or: [{ _id: id }, { slug }] }).exec();
-
-      if (!dish) {
-        throw new Error("Dish not found.");
-      }
-
-      // Fetch special dietary options for the restaurant
-      const existingSpecialDiets = await getSpecialDiets();
-
-      const dishDTO = createDishDTO(dish, existingSpecialDiets);
-
-      return dishDTO;
-    } catch (error) {
-      console.error(error);
-      throw new Error("Failed to fetch dish.");
-    }
-  },
-  ["dish"],
-  { tags: ["dish"], revalidate: 60 * 60 * 24 }
-);
+    await Dish.insertMany(dishesDemo);
+  } catch (error) {
+    console.error(error);
+    throw new Error("Failed to create demo dishes.");
+  }
+};
 
 const updateDish = async ({ id, data }: { id: string | undefined; data: MenuSchema }): Promise<string> => {
   try {
-    await validateRole(["admin", "superadmin"]);
+    const { user } = await validateRole(["admin", "superadmin", "demo"]);
+    const isDemo = user.role === "demo";
 
     if (data.imageFile.size) {
       const uploadResult = await uploadImage(data.imageFile, {
-        folder: `/food-ordering-app/restaurant/menu/${data.course}`,
+        folder: `/food-ordering-app/restaurant${isDemo ? "-demo" : ""}/menu/${data.course}`,
         publicId: data.image.publicId || undefined,
       });
 
@@ -134,8 +193,13 @@ const updateDish = async ({ id, data }: { id: string | undefined; data: MenuSche
     await dbConnect();
 
     const dish = id
-      ? await Dish.findOneAndUpdate({ _id: id }, dbReadyData, { new: true }).exec()
-      : await Dish.create(dbReadyData);
+      ? await Dish.findOneAndUpdate({ _id: id, userId: isDemo ? user.id : undefined }, dbReadyData, {
+          new: true,
+        }).exec()
+      : await Dish.create({
+          ...dbReadyData,
+          ...(isDemo ? { userId: user.id, expiresAt: new Date(Date.now() + 1000 * 60 * 60) } : {}),
+        });
 
     if (!dish) {
       throw new Error("Update failed. Dish not found.");
@@ -156,10 +220,11 @@ const updateDish = async ({ id, data }: { id: string | undefined; data: MenuSche
 
 const deleteDish = async ({ id }: { id: string }): Promise<void> => {
   try {
-    await validateRole(["admin", "superadmin"]);
+    const { user } = await validateRole(["admin", "superadmin", "demo"]);
+    const isDemo = user.role === "demo";
 
     await dbConnect();
-    const dish = await Dish.findOneAndDelete({ _id: id }).exec();
+    const dish = await Dish.findOneAndDelete({ _id: id, userId: isDemo ? user.id : undefined }).exec();
 
     if (!dish) {
       throw new Error("Dish not found.");
@@ -172,4 +237,4 @@ const deleteDish = async ({ id }: { id: string }): Promise<void> => {
   }
 };
 
-export { deleteDish, getDish, getDishes, updateDish, type DishDTO };
+export { createDemoDishes, deleteDish, getDish, getDishes, updateDish, type DishDTO };
